Move mandating two step verification test into shared test file
Also adds a test helper for common managing two sv helpers
